pyproject.toml
[tool.poetry]
name = "hr-worker"
version = "0.1.0"
description = "HR Worker Management System"
authors = ["Your Name <your.email@example.com>"]
package-mode = false  # ← Добавьте это: отключает установку проекта как пакета

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
djangorestframework = "^3.14"
drf-yasg = "^1.21"
django-filter = "^23.1"
openpyxl = "^3.1"

[tool.poetry.group.dev.dependencies]  # ← Замените dev-dependencies на это
pytest = "^7.2"
pytest-django = "^4.5"
black = "^23.1"
flake8 = "^6.0"
pre-commit = "^3.0"
